Moved Greeting view to components package
